Cleanup fmpz_mod_mat.h
